# perf: optimize cursor dataIdx scanning to viewport range only