… dist/uPlot.cjs.js
… dist/uPlot.esm.js
… dist/uPlot.iife.js