@leeoniya 2025-09-10
hey @bbjbc thanks for the PR :)

i'm pretty sure this was an AI-assisted contribution, so here is some feedback, hopefully for a human:

> Reduces CPU usage by 80-95% in typical zoomed scenarios

this part of the code is very very specific to configurations with a non-default `cursor.hover.bias`, and it only executes when the user has hovered a null data point. additionally, it only scans left and right from the cursor position, so for the existing code to scan the entire dataset, the data would have to completely consist of null values. for 99.999% of scenarios these changes have zero effect on performance. and when they do have some effect it would so small that it would be difficult to measure.

it doesnt take much effort for a human looking at the code to realize this. sorry.

i pushed b64596308632d1a9f221b1f81f2398b671c1f053 which addressed the TODO more cleanly.